[ ESP32 ] : ( Issue_1 )-Firmware Crash when socket connection breaks with server running on application. ( Issue_2 )-Sometimes Wroong JSN address being assigned to Buffer which is being used to send over TCP. Actuall Issue not known but added changes as work around. ( Logic_1 ) - Removed check which monitors the exact clock instant before starting scan or operating output, because sometimes, it comes into the scan part 2ms before               scheduled instant and does not perform scan for next consecutive IOS and remain in the same scan state. ( Logic_2 ) - Enabed Firmware dump generation over UART. ( Logic_3 ) - Reduced Task size from 10K to 4K for both TCP RX and TCP TX.

+++ Firmwares/ESP32/Quad64/main/src/commonDeclaration.c
@@ -396,7 +396,7 @@ void delayMicroseconds(uint32_t _delayInuS)
 */
 
 #if defined( SYNC_NATIVE_CLOCK )
-uint64_t clockOffset = 0;
+uint32_t clockOffset = 0;
 bool clockOffsettingDone = false;
 bool clockValueLocked = false;
 uint64_t lockedClockValue = 0;
